single_image_3d/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.2
opencv-python>=4.8.0
# Pin below NumPy 2 to avoid compiled-extension incompatibilities in some wheels
numpy>=1.24.0,<2.0.0
 pillow>=10.0.0
 open3d>=0.17.0
 trimesh>=4.0.0
 scikit-image>=0.21.0
 imageio>=2.34.0
 rich>=13.7.0
